feat: Implement ESP32 Control GUI with serial communication and command panels
- Added main GUI class `ESP32ControlGUI` for controlling the Ghost ESP32 device.
- Implemented serial communication using `pyserial` with a dedicated thread for monitoring incoming data.
- Created command panels for WiFi, Network, BLE, Evil Portal, and Settings operations.
- Added functionality for sending commands to the ESP32 and processing responses.
- Introduced auto-reconnect feature for serial connection.
- Implemented logging and display area for command responses and logs.
- Added utility functions for timestamping and logging messages.
- Created separate threads for sending HTML files as portals and monitoring serial data.
- Set up a main entry point in `main.py` to launch the application.
